TEIIDTOOLS-723 consolidating spi and core
