+++ server/komodo-core/pom.xml
@@ -42,10 +42,6 @@
       <artifactId>teiid-engine</artifactId>
     </dependency>  
     <dependency>
-      <groupId>org.teiid</groupId>
-      <artifactId>komodo-spi</artifactId>
-    </dependency>
-    <dependency>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-data-jpa</artifactId>
     </dependency>
@@ -65,25 +61,4 @@
 	</dependency>    
   </dependencies>
   
-  <build>
-    <testResources>
-      <testResource>
-        <directory>${project.basedir}/src/test/resources</directory>
-        <filtering>true</filtering>
-      </testResource>
-    </testResources>
-
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-resources-plugin</artifactId>
-        <configuration>
-          <nonFilteredFileExtensions>
-            <nonFilteredFileExtension>ddl</nonFilteredFileExtension>
-            <nonFilteredFileExtension>txt</nonFilteredFileExtension>
-          </nonFilteredFileExtensions>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
 </project>
