+++ server/komodo-core/src/main/java/org/komodo/KException.java
