+++ server/komodo-core/src/main/java/org/komodo/StringConstants.java
