+++ server/komodo-core/src/main/java/org/komodo/TeiidSqlConstants.java
