+++ server/komodo-core/src/main/java/org/komodo/WorkspaceManager.java
