+++ server/komodo-core/src/main/java/org/komodo/datavirtualization/DataVirtualization.java
@@ -15,44 +15,69 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.komodo.datavirtualization;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+import org.hibernate.annotations.GenericGenerator;
 import org.komodo.KException;
 import org.komodo.StringConstants;
 
-/**
- * A model of a datavirtualization instance
- */
-public interface DataVirtualization extends Named {
-	
+@Entity
+public class DataVirtualization implements Named {
+
 	static String getServiceVdbName(String name) {
 		return name.toLowerCase() + StringConstants.SERVICE_VDB_SUFFIX;
 	}
-
+	
+	@Id
+	@GeneratedValue(generator = "uuid2")
+    @GenericGenerator(name = "uuid2", strategy = "org.hibernate.id.UUIDGenerator")
+	private String id;
+	@Column(unique=true)
+	private String name;
+	private String description;
+	
+	protected DataVirtualization() {
+	}
+	
+	public DataVirtualization(String name) {
+		this.name = name;
+	}
+	
     /**
      * @return the service VDB name (may be <code>null</code> if not defined)
      */
-    default String getServiceVdbName( ) throws KException {
+    public String getServiceVdbName( ) throws KException {
     	return getServiceVdbName(getName());
     }
-
-    /**
-     * @return the value of the <code>description</code> property (can be empty)
-     */
-    String getDescription( );
-
-    /**
-     * @param newDescription
-     *        the new value of the <code>description</code> property
-     */
-    void setDescription( 
-                         final String newDescription );
-    
-    /**
-     * Get the id of the virtualization
-     */
-    String getId();
-
-	void setName(String name);
-    
+	
+	public String getId() {
+		return id;
+	}
+	
+	public void setId(String id) {
+		this.id = id;
+	}
+	
+	public String getName() {
+		return name;
+	}
+	
+	public void setName(String name) {
+		this.name = name;
+	}
+	
+	public String getDescription() {
+		return description;
+	}
+	
+	public void setDescription(String description) {
+		this.description = description;
+	}
+	
 }
