+++ server/komodo-core/src/main/java/org/komodo/datavirtualization/Named.java
