+++ server/komodo-core/src/main/java/org/komodo/datavirtualization/SourceSchema.java
@@ -18,10 +18,46 @@
 
 package org.komodo.datavirtualization;
 
-public interface SourceSchema extends Named {
+import javax.persistence.Entity;
+import javax.persistence.Id;
 
-	String getDdl();
+@Entity
+public class SourceSchema implements Named {
 
-	String getId();
+	@Id
+	private String id;
+	private String ddl;
+	private String name;
+	
+	protected SourceSchema() {
+	}
+	
+	public SourceSchema(String id) {
+		this.id = id;
+	}
+	
+	public String getDdl() {
+		return ddl;
+	}
+
+	public void setDdl(String ddl) {
+		this.ddl = ddl;
+	}
+	
+	public String getId() {
+		return id;
+	}
+	
+	public void setId(String id) {
+		this.id = id;
+	}
+	
+	public String getName() {
+		return name;
+	}
+	
+	public void setName(String name) {
+		this.name = name;
+	}
 
 }
