+++ server/komodo-core/src/main/java/org/komodo/repository/DataVirtualizationRepository.java
@@ -18,6 +18,7 @@
 
 package org.komodo.repository;
 
+import org.komodo.datavirtualization.DataVirtualization;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
