+++ server/komodo-core/src/main/java/org/komodo/repository/KomodoRepositoryConfiguration.java
@@ -18,6 +18,7 @@
 
 package org.komodo.repository;
 
+import org.komodo.datavirtualization.ViewDefinition;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
@@ -25,7 +26,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
 @EnableJpaRepositories
-@EntityScan
+@EntityScan(basePackageClasses = ViewDefinition.class)
 @ComponentScan
 public class KomodoRepositoryConfiguration {
