--- server/komodo-core/src/main/java/org/komodo/repository/SourceSchema.java
@@ -1,65 +1,0 @@
-/*
- * Copyright Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags and
- * the COPYRIGHT.txt file distributed with this work.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.komodo.repository;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-
-@Entity
-public class SourceSchema implements org.komodo.datavirtualization.SourceSchema {
-
-	@Id
-	private String id;
-	private String ddl;
-	private String name;
-	
-	protected SourceSchema() {
-	}
-	
-	public SourceSchema(String id) {
-		this.id = id;
-	}
-	
-	@Override
-	public String getDdl() {
-		return ddl;
-	}
-
-	public void setDdl(String ddl) {
-		this.ddl = ddl;
-	}
-	
-	@Override
-	public String getId() {
-		return id;
-	}
-	
-	public void setId(String id) {
-		this.id = id;
-	}
-	
-	public String getName() {
-		return name;
-	}
-	
-	public void setName(String name) {
-		this.name = name;
-	}
-
-}
