+++ server/komodo-core/src/main/java/org/komodo/repository/SourceSchemaRepository.java
@@ -20,6 +20,7 @@ package org.komodo.repository;
 
 import java.util.List;
 
+import org.komodo.datavirtualization.SourceSchema;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
