+++ server/komodo-core/src/main/java/org/komodo/utils/KLog.java
