+++ server/komodo-core/src/main/java/org/komodo/utils/StringNameValidator.java
