+++ server/komodo-core/src/main/java/org/komodo/utils/StringUtils.java
