+++ server/komodo-core/src/main/java/org/komodo/utils/TextI18n.java
