+++ server/komodo-core/src/main/resources/org/komodo/messages.properties
@@ -1,0 +1,2 @@
+SPI.valueCannotBeEmpty = Value {0} cannot be null or zero length
+SPI.valueCannotBeNull = Value {0} cannot be null
