+++ server/komodo-core/src/test/java/org/komodo/utils/FileUtilsTest.java
