+++ server/komodo-core/src/test/java/org/komodo/utils/TestKLog.java
