--- server/komodo-spi/src/main/resources/org/komodo/messages.properties
@@ -1,10 +1,0 @@
-SPI.valueCannotBeEmpty = Value {0} cannot be null or zero length
-SPI.valueCannotBeNull = Value {0} cannot be null
-
-SPI.invalidTargetTypeForGetTeiidMethod = Invalid target type {0} for execution event. Method getServer() expected type {1}.
-SPI.invalidTargetTypeForGetDataSourceMethod = Invalid target type {0} for execution event. Method getDataSource() expected type {1}.
-SPI.invalidTargetTypeForGetTranslatorMethod = Invalid target type {0} for execution event. Method getTranslator() expected type {1}.
-SPI.invalidTargetTypeForGetUpdatedTeiidMethod = Invalid target type {0} for execution event. Method getUpdatedServer() expected type {1}.
-
-
-OutcomeFactory.OK = OK
