+++ server/pom.xml
@@ -186,7 +186,6 @@
 
 
   <modules>
-    <module>komodo-spi</module>
     <module>komodo-core</module>
     <module>komodo-server</module>
   </modules>
@@ -251,11 +250,6 @@
       <!-- Komodo subprojects -->
       <dependency>
         <groupId>org.teiid</groupId>
-        <artifactId>komodo-spi</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.teiid</groupId>
         <artifactId>komodo-metadata</artifactId>
         <version>${project.version}</version>
       </dependency>
@@ -268,13 +262,6 @@
         <groupId>org.teiid</groupId>
         <artifactId>komodo-server</artifactId>
         <version>${project.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.teiid</groupId>
-        <artifactId>komodo-spi</artifactId>
-        <version>${project.version}</version>
-        <type>test-jar</type>
-        <scope>test</scope>
       </dependency>
       <dependency>
         <groupId>org.teiid</groupId>
